zebra: start detangling rtadv & irdp
Replace some cross-dependencies with hooks & move bits to where they
belong.
