+++ zebra/rtadv.c
@@ -62,6 +62,10 @@ extern struct zebra_privs_t zserv_privs;
 #define ALLNODE   "ff02::1"
 #define ALLROUTER "ff02::2"
 
+/* Order is intentional.  Matches RFC4191.  This array is also used for
+   command matching, so only modify with care. */
+const char *rtadv_pref_strs[] = {"medium", "high", "INVALID", "low", 0};
+
 enum rtadv_event {
 	RTADV_START,
 	RTADV_STOP,
@@ -1456,9 +1460,76 @@ DEFUN (no_ipv6_nd_mtu,
 	return CMD_SUCCESS;
 }
 
+/* Dump interface ND information to vty. */
+static int nd_dump_vty(struct vty *vty, struct interface *ifp)
+{
+	struct zebra_if *zif;
+	struct rtadvconf *rtadv;
+	int interval;
+
+	zif = (struct zebra_if *)ifp->info;
+	rtadv = &zif->rtadv;
+
+	if (rtadv->AdvSendAdvertisements) {
+		vty_out(vty,
+			"  ND advertised reachable time is %d milliseconds\n",
+			rtadv->AdvReachableTime);
+		vty_out(vty,
+			"  ND advertised retransmit interval is %d milliseconds\n",
+			rtadv->AdvRetransTimer);
+		vty_out(vty, "  ND router advertisements sent: %d rcvd: %d\n",
+			zif->ra_sent, zif->ra_rcvd);
+		interval = rtadv->MaxRtrAdvInterval;
+		if (interval % 1000)
+			vty_out(vty,
+				"  ND router advertisements are sent every "
+				"%d milliseconds\n",
+				interval);
+		else
+			vty_out(vty,
+				"  ND router advertisements are sent every "
+				"%d seconds\n",
+				interval / 1000);
+		if (rtadv->AdvDefaultLifetime != -1)
+			vty_out(vty,
+				"  ND router advertisements live for %d seconds\n",
+				rtadv->AdvDefaultLifetime);
+		else
+			vty_out(vty,
+				"  ND router advertisements lifetime tracks ra-interval\n");
+		vty_out(vty,
+			"  ND router advertisement default router preference is "
+			"%s\n",
+			rtadv_pref_strs[rtadv->DefaultPreference]);
+		if (rtadv->AdvManagedFlag)
+			vty_out(vty,
+				"  Hosts use DHCP to obtain routable addresses.\n");
+		else
+			vty_out(vty,
+				"  Hosts use stateless autoconfig for addresses.\n");
+		if (rtadv->AdvHomeAgentFlag) {
+			vty_out(vty,
+				"  ND router advertisements with Home Agent flag bit set.\n");
+			if (rtadv->HomeAgentLifetime != -1)
+				vty_out(vty,
+					"  Home Agent lifetime is %u seconds\n",
+					rtadv->HomeAgentLifetime);
+			else
+				vty_out(vty,
+					"  Home Agent lifetime tracks ra-lifetime\n");
+			vty_out(vty, "  Home Agent preference is %u\n",
+				rtadv->HomeAgentPreference);
+		}
+		if (rtadv->AdvIntervalOption)
+			vty_out(vty,
+				"  ND router advertisements with Adv. Interval option.\n");
+	}
+	return 0;
+}
+
 
 /* Write configuration about router advertisement. */
-void rtadv_config_write(struct vty *vty, struct interface *ifp)
+static int rtadv_config_write(struct vty *vty, struct interface *ifp)
 {
 	struct zebra_if *zif;
 	struct listnode *node;
@@ -1539,6 +1610,7 @@ void rtadv_config_write(struct vty *vty, struct interface *ifp)
 			vty_out(vty, " router-address");
 		vty_out(vty, "\n");
 	}
+	return 0;
 }
 
 
@@ -1600,6 +1672,9 @@ void rtadv_terminate(struct zebra_ns *zns)
 
 void rtadv_cmd_init(void)
 {
+	hook_register(zebra_if_extra_info, nd_dump_vty);
+	hook_register(zebra_if_config_wr, rtadv_config_write);
+
 	install_element(INTERFACE_NODE, &ipv6_nd_suppress_ra_cmd);
 	install_element(INTERFACE_NODE, &no_ipv6_nd_suppress_ra_cmd);
 	install_element(INTERFACE_NODE, &ipv6_nd_ra_interval_cmd);
